+++ vite-plugin-ssr/utils/isPromise.ts
@@ -1,8 +1,4 @@
-import { hasProp } from './hasProp'
 import { isCallable } from './isCallable'
-
-export { isPromise }
-
-function isPromise(thing: unknown): thing is Promise<unknown> {
-  return hasProp(thing, 'then') && isCallable(thing.then)
+export function isPromise(val: unknown): val is Promise<unknown> {
+  return typeof val === 'object' && val !== null && 'then' in val && isCallable((val as Record<string, unknown>).then)
 }